// src/hooks/usePickCep.types.ts
export type aboutType = {
  avatar: string;
  bairro: string;
  cep: string;
  complemento: string;
  ddd: string;
  gia: string;
  ibge: string;
  localidade: string;
  logradouro: string;
  nome: string;
  apelido: string;
  siafi: string;
  uf: string;
};

export type formacaoTypes = {
  escola: string;
  curso: string;
  tempo: string;
  inicio: string;
};

export type xpTypes = {
  empresa: string;
  cargo: string;
  descricao: string[];
  stacks: string[];
};